add type to get it right

+++ org/rfid/libdanrfid/DDMTag.java
@@ -2,8 +2,8 @@ package org.rfid.libdanrfid;
 
 public class DDMTag extends DDMData {
 
-	public static final byte AFI_ON=0x07;
-	public static final byte AFI_OFF=0xC2;
+	public static final byte AFI_ON=(byte)0x07;
+	public static final byte AFI_OFF=(byte)0xC2;
 
 	private byte afi=0;
 	public boolean isTainted=true; // we have not read all - or have made changes yet unwritten
